# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for managing Akun resources",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sub_akun": {
            "get": {
                "tags": [
                    "SubAkun"
                ],
                "summary": "Get list of SubAkun",
                "description": "Display a listing of SubAkun.",
                "operationId": "02855f78008af0670c4d9e3edcd70c64",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubAkun"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SubAkun"
                ],
                "summary": "Create a new SubAkun",
                "description": "Store a new SubAkun.",
                "operationId": "3e0f28f3b9e6005146994742cfed1bc3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "id_akun": {
                                        "type": "string",
                                        "example": "akun_123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "SubAkun created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubAkun"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/sub_akun/{id}": {
            "get": {
                "tags": [
                    "SubAkun"
                ],
                "summary": "Get a SubAkun by ID",
                "description": "Display a SubAkun by ID.",
                "operationId": "b62c27feec5af21315cb09f59670f7d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubAkun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SubAkun not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "SubAkun"
                ],
                "summary": "Update a SubAkun by ID",
                "description": "Update an existing SubAkun.",
                "operationId": "7c5461a670f63b6cc4605abb74f07ca9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "kode": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "id_akun": {
                                        "type": "string",
                                        "example": "akun_456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SubAkun updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubAkun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SubAkun not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SubAkun"
                ],
                "summary": "Delete a SubAkun by ID",
                "description": "Delete a SubAkun.",
                "operationId": "7120ac0088cb476f94c2c7f6d2ac3657",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "SubAkun deleted"
                    },
                    "404": {
                        "description": "SubAkun not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubAkun": {
                "properties": {
                    "id_sub_akun": {
                        "type": "string",
                        "example": "sub_akun_123"
                    },
                    "id_akun": {
                        "type": "string",
                        "example": "akun_123"
                    },
                    "kode": {
                        "type": "integer",
                        "example": 101
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}